Move and split up weather.py.
This is getting out of hand, and I'm about to make it worse.
