+++ game/theater/controlpoint.py
@@ -75,7 +75,7 @@ from ..db import Database
 from ..dcs.aircrafttype import AircraftType
 from ..dcs.groundunittype import GroundUnitType
 from ..utils import nautical_miles
-from ..weather import Conditions
+from ..weather.conditions import Conditions
 
 if TYPE_CHECKING:
     from game import Game
