+++ qt_ui/widgets/QConditionsWidget.py
@@ -15,7 +15,7 @@ import qt_ui.uiconstants as CONST
 from game.sim.gameupdateevents import GameUpdateEvents
 from game.timeofday import TimeOfDay
 from game.utils import mps
-from game.weather import Conditions
+from game.weather.conditions import Conditions
 from qt_ui.simcontroller import SimController
